refactor: Integrate features into a class
